+++ integration-tests/api/__tests__/batch-jobs/api.js
@@ -43,6 +43,13 @@ const setupJobDb = async (dbConnection) => {
       status: "awaiting_confirmation",
       created_by: "member-user",
     })
+    await simpleBatchJobFactory(dbConnection, {
+      id: "job_5",
+      type: "product-export",
+      status: "completed",
+      completed_at: "2022-06-27T22:00:00.000Z",
+      created_by: "admin_user",
+    })
   } catch (err) {
     console.log(err)
     throw err
@@ -56,7 +63,7 @@ describe("/admin/batch-jobs", () => {
   beforeAll(async () => {
     const cwd = path.resolve(path.join(__dirname, "..", ".."))
     dbConnection = await initDb({ cwd })
-    medusaProcess = await setupServer({ cwd, verbose: false })
+    medusaProcess = await setupServer({ cwd })
   })
 
   afterAll(async () => {
@@ -81,23 +88,65 @@ describe("/admin/batch-jobs", () => {
       const response = await api.get("/admin/batch-jobs", adminReqConfig)
 
       expect(response.status).toEqual(200)
+      expect(response.data.batch_jobs.length).toEqual(4)
+      expect(response.data).toMatchSnapshot({
+        batch_jobs: [
+          {
+            id: "job_5",
+            created_at: expect.any(String),
+            updated_at: expect.any(String),
+            created_by: "admin_user",
+          },
+          {
+            id: "job_3",
+            created_at: expect.any(String),
+            updated_at: expect.any(String),
+            created_by: "admin_user",
+          },
+          {
+            id: "job_2",
+            created_at: expect.any(String),
+            updated_at: expect.any(String),
+            created_by: "admin_user",
+          },
+          {
+            id: "job_1",
+            created_at: expect.any(String),
+            updated_at: expect.any(String),
+            created_by: "admin_user",
+          },
+        ],
+      })
+    })
+
+    it("lists batch jobs created by the user and where completed_at is null ", async () => {
+      const api = useApi()
+      const response = await api.get(
+        "/admin/batch-jobs?completed_at=null",
+        adminReqConfig
+      )
+
+      expect(response.status).toEqual(200)
       expect(response.data.batch_jobs.length).toEqual(3)
       expect(response.data).toMatchSnapshot({
         batch_jobs: [
           {
+            id: "job_3",
             created_at: expect.any(String),
             updated_at: expect.any(String),
-            created_by: "admin_user"
+            created_by: "admin_user",
           },
           {
+            id: "job_2",
             created_at: expect.any(String),
             updated_at: expect.any(String),
-            created_by: "admin_user"
+            created_by: "admin_user",
           },
           {
+            id: "job_1",
             created_at: expect.any(String),
             updated_at: expect.any(String),
-            created_by: "admin_user"
+            created_by: "admin_user",
           },
         ],
       })
@@ -119,23 +168,24 @@ describe("/admin/batch-jobs", () => {
       const response = await api.get("/admin/batch-jobs/job_1", adminReqConfig)
 
       expect(response.status).toEqual(200)
-      expect(response.data.batch_job).toEqual(expect.objectContaining({
-        created_at: expect.any(String),
-        updated_at: expect.any(String),
-        created_by: "admin_user"
-      }))
+      expect(response.data.batch_job).toEqual(
+        expect.objectContaining({
+          created_at: expect.any(String),
+          updated_at: expect.any(String),
+          created_by: "admin_user",
+        })
+      )
     })
 
     it("should fail on batch job created by other user", async () => {
       const api = useApi()
-      await api.get("/admin/batch-jobs/job_4", adminReqConfig)
-        .catch((err) => {
-          expect(err.response.status).toEqual(400)
-          expect(err.response.data.type).toEqual("not_allowed")
-          expect(err.response.data.message).toEqual(
-            "Cannot access a batch job that does not belong to the logged in user"
-          )
-        })
+      await api.get("/admin/batch-jobs/job_4", adminReqConfig).catch((err) => {
+        expect(err.response.status).toEqual(400)
+        expect(err.response.data.type).toEqual("not_allowed")
+        expect(err.response.data.message).toEqual(
+          "Cannot access a batch job that does not belong to the logged in user"
+        )
+      })
     })
   })
 
@@ -149,7 +199,7 @@ describe("/admin/batch-jobs", () => {
       await db.teardown()
     })
 
-    it("Creates a batch job", async() => {
+    it("Creates a batch job", async () => {
       const api = useApi()
 
       const response = await api.post(
@@ -215,7 +265,7 @@ describe("/admin/batch-jobs", () => {
       }
     })
 
-    afterEach(async() => {
+    afterEach(async () => {
       const db = useDb()
       await db.teardown()
     })