+++ integration-tests/helpers/use-template-db.js
@@ -31,8 +31,26 @@ class DatabaseFactory {
           `@medusajs`,
           `medusa`,
           `dist`,
-          `migrations`
+          `migrations`,
+          `*.js`
         )
+      )
+
+      const { getEnabledMigrations } = require(path.join(
+        cwd,
+        `node_modules`,
+        `@medusajs`,
+        `medusa`,
+        `dist`,
+        `commands`,
+        `utils`,
+        `get-migrations`
+      ))
+
+      // filter migrations to only include those that dont have feature flags
+      const enabledMigrations = await getEnabledMigrations(
+        [migrationDir],
+        (flag) => false
       )
 
       await dropDatabase(
@@ -51,7 +69,7 @@ class DatabaseFactory {
         type: "postgres",
         name: "templateConnection",
         url: `${DB_URL}/${this.templateDbName}`,
-        migrations: [`${migrationDir}/*.js`],
+        migrations: enabledMigrations,
       })
 
       await templateDbConnection.runMigrations()
@@ -92,7 +110,7 @@ class DatabaseFactory {
   }
 
   async destroy() {
-    let connection = await this.getMasterConnection()
+    const connection = await this.getMasterConnection()
 
     await connection.query(`DROP DATABASE IF EXISTS "${this.templateDbName}";`)
     await connection.close()