+++ packages/medusa-plugin-add-ons/src/api/middlewares/index.js
@@ -1,5 +1,0 @@
-import { default as wrap } from "./await-middleware"
-
-export default {
-  wrap,
-}