+++ packages/medusa/src/api/routes/admin/collections/create-collection.ts
@@ -1,6 +1,7 @@
 import { IsNotEmpty, IsObject, IsOptional, IsString } from "class-validator"
 import ProductCollectionService from "../../../../services/product-collection"
-import { validator } from "../../../../utils/validator"
+import { Request, Response } from "express"
+
 /**
  * @oas [post] /collections
  * operationId: "PostCollections"
@@ -35,14 +36,14 @@ import { validator } from "../../../../utils/validator"
  *            collection:
  *              $ref: "#/components/schemas/product_collection"
  */
-export default async (req, res) => {
-  const validated = await validator(AdminPostCollectionsReq, req.body)
+export default async (req: Request, res: Response) => {
+  const { validatedBody } = req
 
   const productCollectionService: ProductCollectionService = req.scope.resolve(
     "productCollectionService"
   )
 
-  const created = await productCollectionService.create(validated)
+  const created = await productCollectionService.create(validatedBody)
   const collection = await productCollectionService.retrieve(created.id)
 
   res.status(200).json({ collection })