+++ packages/medusa/src/api/routes/admin/customer-groups/add-customers-batch.ts
@@ -3,6 +3,7 @@ import { ValidateNested } from "class-validator"
 import { CustomerGroupService } from "../../../../services"
 import { CustomerGroupsBatchCustomer } from "../../../../types/customer-groups"
 import { validator } from "../../../../utils/validator"
+import { Request, Response } from "express"
 
 /**
  * @oas [post] /customer-groups/{id}/customers/batch
@@ -40,7 +41,7 @@ import { validator } from "../../../../utils/validator"
  *               $ref: "#/components/schemas/customer_group"
  */
 
-export default async (req, res) => {
+export default async (req: Request, res: Response) => {
   const { id } = req.params
   const validated = await validator(
     AdminPostCustomerGroupsGroupCustomersBatchReq,