+++ packages/medusa/src/api/routes/admin/customer-groups/index.ts
@@ -1,38 +1,53 @@
 import { Router } from "express"
 import { CustomerGroup } from "../../../.."
 import { DeleteResponse, PaginatedResponse } from "../../../../types/common"
-import middlewares from "../../../middlewares"
+import middlewares, { transformQuery } from "../../../middlewares"
+import { AdminGetCustomerGroupsGroupParams } from "./get-customer-group"
+import { AdminGetCustomerGroupsParams } from "./list-customer-groups"
 
 const route = Router()
 
 export default (app) => {
   app.use("/customer-groups", route)
 
-  route.get("/", middlewares.wrap(require("./list-customer-groups").default))
-  route.get("/:id", middlewares.wrap(require("./get-customer-group").default))
   route.post("/", middlewares.wrap(require("./create-customer-group").default))
-  route.post(
-    "/:id/customers/batch",
-    middlewares.wrap(require("./add-customers-batch").default)
-  )
-  route.delete(
-    "/:id/customers/batch",
-    middlewares.wrap(require("./delete-customers-batch").default)
+  route.get(
+    "/",
+    transformQuery(AdminGetCustomerGroupsParams, {
+      defaultRelations: defaultAdminCustomerGroupsRelations,
+      isList: true,
+    }),
+    middlewares.wrap(require("./list-customer-groups").default)
   )
 
-  route.delete(
-    "/:id",
+  const customerGroupRouter = Router({ mergeParams: true })
+  route.use("/:id", customerGroupRouter)
+  customerGroupRouter.get(
+    "/",
+    transformQuery(AdminGetCustomerGroupsGroupParams, {
+      defaultRelations: defaultAdminCustomerGroupsRelations,
+    }),
+    middlewares.wrap(require("./get-customer-group").default)
+  )
+  customerGroupRouter.delete(
+    "/",
     middlewares.wrap(require("./delete-customer-group").default)
   )
-
-  route.get(
-    "/:id/customers",
+  customerGroupRouter.post(
+    "/",
+    middlewares.wrap(require("./update-customer-group").default)
+  )
+  customerGroupRouter.get(
+    "/customers",
     middlewares.wrap(require("./get-customer-group-customers").default)
   )
-
-  route.post(
-    "/:id",
-    middlewares.wrap(require("./update-customer-group").default)
+  customerGroupRouter.post(
+    "/customers/batch",
+    middlewares.wrap(require("./add-customers-batch").default)
+  )
+  customerGroupRouter.delete(
+    "/customers/batch",
+    middlewares.wrap(require("./delete-customers-batch").default)
   )
 
   return app