+++ packages/medusa/src/api/routes/admin/uploads/create-upload.ts
@@ -51,7 +51,7 @@ export default async (req, res) => {
   }
 }
 
-export class IAdminPostUploadsFile {
+export class IAdminPostUploadsFileReq {
   originalName: string
   path: string
 }