+++ packages/medusa/src/api/routes/admin/uploads/get-download-url.ts
@@ -1,12 +1,11 @@
 import { IsString } from "class-validator"
 import { AbstractFileService } from "../../../../interfaces"
-import { validator } from "../../../../utils/validator"
 
 /**
- * [get] /uploads
- * operationId: "GetUploadsFileDownloadUrl"
- * summary: "Gets a presigned download url for a file"
- * description: "Gets a presigned download url for a file"
+ * [post] /uploads/download-url
+ * operationId: "PostUploadsDownloadUrl"
+ * summary: "Creates a presigned download url for a file"
+ * description: "Creates a presigned download url for a file"
  * x-authenticated: true
  * requestBody:
  *   content:
@@ -26,18 +25,16 @@ import { validator } from "../../../../utils/validator"
  *     description: OK
  */
 export default async (req, res) => {
-  const validated = await validator(AdminGetUploadsFileDownloadUrlReq, req.body)
-
   const fileService: AbstractFileService<any> = req.scope.resolve("fileService")
 
   const url = await fileService.getPresignedDownloadUrl({
-    fileKey: validated.file_key,
+    fileKey: (req.validatedBody as AdminPostUploadsDownloadUrlReq).file_key,
   })
 
   res.status(200).send({ download_url: url })
 }
 
-class AdminGetUploadsFileDownloadUrlReq {
+export class AdminPostUploadsDownloadUrlReq {
   @IsString()
   file_key: string
 }