+++ packages/medusa/src/types/pricing.ts
@@ -1,4 +1,4 @@
-import { MoneyAmount, ProductVariant, Product } from "../models"
+import { MoneyAmount, ProductVariant, Product, ShippingOption } from "../models"
 import { TaxServiceRate } from "./tax-service"
 import { PriceSelectionContext } from "../interfaces/price-selection-strategy"
 
@@ -22,6 +22,14 @@ export type PricingContext = {
   automatic_taxes: boolean
   tax_rate: number | null
 }
+
+export type ShippingOptionPricing = {
+  price_incl_tax: number | null
+  tax_rates: TaxServiceRate[] | null
+}
+
+export type PricedShippingOption = Partial<ShippingOption> &
+  ShippingOptionPricing
 
 export type PricedVariant = Partial<ProductVariant> & ProductVariantPricing