+++ packages/medusa-plugin-twilio-sms/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.3](https://github.com/medusajs/medusa/compare/medusa-plugin-twilio-sms@1.2.2...medusa-plugin-twilio-sms@1.2.3) (2022-07-05)
+
+**Note:** Version bump only for package medusa-plugin-twilio-sms
+
+
+
+
+
 ## [1.2.2](https://github.com/medusajs/medusa/compare/medusa-plugin-twilio-sms@1.2.0...medusa-plugin-twilio-sms@1.2.2) (2022-06-19)
 
 **Note:** Version bump only for package medusa-plugin-twilio-sms